# saga-pattern-example/order-service/src/main/java/com/huongdanjava/systemdesign/sagapattern/orderservice/web/OrdersApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.23.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.huongdanjava.systemdesign.sagapattern.orderservice.web;

import com.huongdanjava.systemdesign.sagapattern.orderservice.web.dto.AddNewOrderResponse;
import com.huongdanjava.systemdesign.sagapattern.orderservice.web.dto.InlineObject;
import org.springframework.lang.Nullable;
import com.huongdanjava.systemdesign.sagapattern.orderservice.web.dto.Order;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import jakarta.validation.Valid;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-04T12:52:53.676943+07:00[Asia/Ho_Chi_Minh]", comments = "Generator version: 7.23.0")
@Validated
@Tag(name = "orders", description = "the orders API")
public interface OrdersApi {

    default OrdersApiDelegate getDelegate() {
        return new OrdersApiDelegate() {};
    }

    String PATH_ADD_NEW_ORDER = "/orders";
    /**
     * POST /orders : Add a new order
     *
     * @param order  (optional)
     * @return Add new order successfully (status code 200)
     *         or Bad Request (status code 400)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "addNewOrder",
        summary = "Add a new order",
        responses = {
            @ApiResponse(responseCode = "200", description = "Add new order successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AddNewOrderResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = InlineObject.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = InlineObject.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = OrdersApi.PATH_ADD_NEW_ORDER,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<AddNewOrderResponse> addNewOrder(
        @Parameter(name = "Order", description = "") @Valid @RequestBody(required = false) @Nullable Order order
    ) {
        return getDelegate().addNewOrder(order);
    }

}